EndpointDescriptor: interface and Default implementation
