+++ model/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/model/descriptor/EndpointDescriptor.java
@@ -14,116 +14,23 @@
  */
 package de.fraunhofer.iosb.ilt.faaast.service.model.descriptor;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import io.adminshell.aas.v3.model.builder.ExtendableBuilder;
 import java.io.Serializable;
-import java.util.Objects;
 
 
 /**
  * Registry Descriptor for Endpoint.
  */
-public class EndpointDescriptor implements Serializable {
+public interface EndpointDescriptor extends Serializable {
 
-    @JsonIgnore
-    private String id;
-    private String interfaceInformation;
-    private ProtocolInformationDescriptor protocolInformation;
-
-    public EndpointDescriptor() {
-
-        id = null;
-        interfaceInformation = null;
-        protocolInformation = null;
-    }
+    public String getInterfaceInformation();
 
 
-    public String getId() {
-        return id;
-    }
+    public void setInterfaceInformation(String interfaceInformation);
 
 
-    public void setId(String id) {
-        this.id = id;
-    }
+    public ProtocolInformationDescriptor getProtocolInformation();
 
 
-    public String getInterfaceInformation() {
-        return interfaceInformation;
-    }
+    public void setProtocolInformation(ProtocolInformationDescriptor protocolInformation);
 
-
-    public void setInterfaceInformation(String interfaceInformation) {
-        this.interfaceInformation = interfaceInformation;
-    }
-
-
-    public ProtocolInformationDescriptor getProtocolInformation() {
-        return protocolInformation;
-    }
-
-
-    public void setProtocolInformation(ProtocolInformationDescriptor protocolInformation) {
-        this.protocolInformation = protocolInformation;
-    }
-
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        EndpointDescriptor that = (EndpointDescriptor) o;
-        return Objects.equals(id, that.id)
-                && Objects.equals(interfaceInformation, that.interfaceInformation)
-                && Objects.equals(protocolInformation, that.protocolInformation);
-    }
-
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, interfaceInformation, protocolInformation);
-    }
-
-
-    public static Builder builder() {
-        return new Builder();
-    }
-
-    public abstract static class AbstractBuilder<T extends EndpointDescriptor, B extends AbstractBuilder<T, B>> extends ExtendableBuilder<T, B> {
-
-        public B id(String value) {
-            getBuildingInstance().setId(value);
-            return getSelf();
-        }
-
-
-        public B interfaceInformation(String value) {
-            getBuildingInstance().setInterfaceInformation(value);
-            return getSelf();
-        }
-
-
-        public B protocolInformation(ProtocolInformationDescriptor value) {
-            getBuildingInstance().setProtocolInformation(value);
-            return getSelf();
-        }
-    }
-
-    public static class Builder extends AbstractBuilder<EndpointDescriptor, Builder> {
-
-        @Override
-        protected Builder getSelf() {
-            return this;
-        }
-
-
-        @Override
-        protected EndpointDescriptor newBuildingInstance() {
-            return new EndpointDescriptor();
-        }
-    }
 }
